Fix panel for mobile
